requirements.txt
Django==2.0.2
djangorestframework==3.7.7
django-filter==2.0.0.dev1
